changing comment from user route to api route
