+++ server/src/controllers/user.controller.js
@@ -6,7 +6,6 @@ import { Task, TaskParticipant } from "../models/task.models.js";
 import { Profile } from "../models/user.models.js";
 import { GENDER, YEAR } from "../models/user.models.js";
 import { addNotification } from './notification.controller.js';
-import { Comment } from '../models/comment.model.js';
 
 
 export const profile = async (req, res, next) => {
@@ -104,20 +103,6 @@ export const task = async (req, res, next) => {
         return res.status(500).json({ error: "Internal server error at user events" });
     }
 };
-
-export const comment = async (req, res, next) => {
-    const id = req.user.id;
-    const { eventId } = req.query;
-    if (!eventId) return res.status(404).json({error: "eventId not included in the query"});
-    try{
-        const profile = await Profile.findOne({userid: id}).select('_id');
-        const result = await Comment.find({"eventId": eventId}).populate('userId', "_id name profilePic");
-        return res.status(200).json(result);
-
-    }catch(err){
-        return res.status(500).json({ error: "Internal server error at Comment" });
-    }
-}
 
 export const event = async (req, res, next) => {
     const id = req.user.id;
