+++ server/src/routes/api.routes.js
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import { getevents, getnews, eventinfo, getnewss } from '../controllers/api.controller.js';
+import { getevents, getnews, eventinfo, comment } from '../controllers/api.controller.js';
 
 const ApiRouter = Router();
 
@@ -8,6 +8,7 @@ ApiRouter.get('/events', getevents);
 // /api/news?date=<DATE>&AdminName=<NAME> 
 ApiRouter.get('/news/:num', getnews);
 
+UserRouter.get("/comments", comment);
 
 ApiRouter.get('/event/info', eventinfo);
