+++ server/src/routes/user.routes.js
@@ -14,7 +14,6 @@ import {
     comment
 } from '../controllers/user.controller.js';
 import { todolist, todo, todoadd, tododelete, tododone } from '../controllers/todo.controller.js';
-import { User } from '../models/user.models.js';
 
 
 const UserRouter = Router();
@@ -31,7 +30,6 @@ UserRouter.post('/event/register', registerInEvent);
 UserRouter.post('/event/task/register', registerInTask);
 
 UserRouter.get("/task", task);
-UserRouter.get("/comments", comment);
 
 UserRouter.get('/todo', todolist);
 UserRouter.get('/todo/:todoId', todo);
